+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Modified `rm-subscription-locations` and
+  `rm-network-location-watcher-relationships` steps to not create
+  `azure_location` entities explicitely but to create mapped relationships with
+  `skipTargetCreation: false`.
+
 ## [5.36.2] - 2022-03-09
 
 ### Fixed